+++ packages/terra-core-docs/src/terra-dev-site/test/toggle/AnimatedToggle.test.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Toggle from '../../../Toggle';
+import Toggle from 'terra-toggle';
 
 const AnimatedToggle = () => {
   const [isOpen, setIsOpen] = useState(false);