+++ packages/terra-toggle/tests/wdio/toggle-spec.js
@@ -1,6 +1,6 @@
 Terra.describeViewports('Toggle', ['medium'], () => {
   it('should expand the toggle', () => {
-    browser.url('/raw/tests/terra-toggle/toggle/default-toggle');
+    browser.url('/raw/tests/cerner-terra-core-docs/toggle/default-toggle');
     $('#trigger-toggle').click();
     Terra.validates.element('expanded toggle');
   });
@@ -11,13 +11,13 @@ Terra.describeViewports('Toggle', ['medium'], () => {
   });
 
   it('should display an opened toggle', () => {
-    browser.url('/raw/tests/terra-toggle/toggle/open-toggle');
+    browser.url('/raw/tests/cerner-terra-core-docs/toggle/open-toggle');
     $('#root').moveTo({ xOffset: 0, yOffset: 0 });
     Terra.validates.element('opened toggle');
   });
 
   it('should disable focusable elements when closed', () => {
-    browser.url('/raw/tests/terra-toggle/toggle/animated-toggle');
+    browser.url('/raw/tests/cerner-terra-core-docs/toggle/animated-toggle');
     expect($('#toggle').getCSSProperty('visibility').value).toBe('hidden');
   });